Organize code a bit more
