+++ packages/shared/src/lib/core/layer-2/actions/index.ts
@@ -2,6 +2,7 @@ export * from './buildEvmTransactionData'
 export * from './buildUnwrapAssetParameters'
 export * from './buildUnwrapAssetTargetAddress'
 export * from './canAccountMakeEvmTransaction'
+export * from './checkForUntrackedTokens'
 export * from './fetchL2BalanceForAccount'
 export * from './getGasFeeForLayer1ToLayer2Transaction'
 export * from './generateAndStoreEvmAddressForAccounts'
