+++ packages/shared/src/lib/core/layer-2/types/layer-2-account-balance.type.ts
@@ -2,6 +2,6 @@ import { NetworkId } from '@core/network/types'
 
 export type Layer2AccountBalance = {
     [networkId in NetworkId]?: {
-        [tokenId: string]: number | bigint
+        [tokenId: string]: number
     }
 }
